refactor: extract devModeParser types to separate file
- Add src/utils/static/devModeParser-types.ts containing all type definitions:
  ConfigSource, ModeStatus, ConfigResult, HierarchyKey,
  ModeEvaluationOptions, ConfigSingleton.
- Update src/utils/static/devModeParser.ts to import types from the new file
  and re-export them at the bottom.
- No runtime behavior changes; types are moved to support separate *-types.ts convention.
